+++ internal/domain/entities/suscription.go
@@ -1,9 +1,9 @@
 package entities
 
 type Subscription struct {
-	ID        int64
-	UserID    int64
-	PlanID    int64
+	ID        int
+	UserID    int
+	PlanID    int
 	Status    string
 	StartedAt string
 	ExpiresAt string
