+++ internal/infrastructure/security/jwt.go
@@ -13,18 +13,18 @@ var (
 )
 
 type Claims struct {
-	UserID int64 `json:"user_id"`
+	UserID int `json:"user_id"`
 	jwt.RegisteredClaims
 }
 
 type VerificationClaims struct {
-	UserID                int64     `json:"user_id"`
+	UserID                int     `json:"user_id"`
 	Email                 string    `json:"email"`
 	VerificationExpiresAt time.Time `json:"verification_expires_at"`
 	jwt.RegisteredClaims
 }
 
-func CreateAccessToken(userID int64) (string, error) {
+func CreateAccessToken(userID int) (string, error) {
 	secret := os.Getenv("JWT_ACCESS_SECRET")
 	if secret == "" {
 		return "", errors.New("JWT_ACCESS_SECRET not set")
@@ -42,7 +42,7 @@ func CreateAccessToken(userID int64) (string, error) {
 	return token.SignedString([]byte(secret))
 }
 
-func CreateRefreshToken(userID int64) (string, error) {
+func CreateRefreshToken(userID int) (string, error) {
 	secret := os.Getenv("JWT_REFRESH_SECRET")
 	if secret == "" {
 		return "", errors.New("JWT_REFRESH_SECRET not set")
@@ -60,7 +60,7 @@ func CreateRefreshToken(userID int64) (string, error) {
 	return token.SignedString([]byte(secret))
 }
 
-func CreateVerificationToken(userID int64, email string, verificationExpiresAt time.Time) (string, error) {
+func CreateVerificationToken(userID int, email string, verificationExpiresAt time.Time) (string, error) {
 	secret := os.Getenv("JWT_VERIFICATION_SECRET") 
 	if secret == "" {
 		return "", errors.New("JWT_VERIFICATION_SECRET not set")
@@ -79,15 +79,15 @@ func CreateVerificationToken(userID int64, email string, verificationExpiresAt t
 	return token.SignedString([]byte(secret))
 }
 
-func ValidateAccessToken(tokenStr string) (int64, error) {
+func ValidateAccessToken(tokenStr string) (int, error) {
 	return validateToken(tokenStr, os.Getenv("JWT_ACCESS_SECRET"))
 }
 
-func ValidateRefreshToken(tokenStr string) (int64, error) {
+func ValidateRefreshToken(tokenStr string) (int, error) {
 	return validateToken(tokenStr, os.Getenv("JWT_REFRESH_SECRET"))
 }
 
-func ValidateVerificationToken(tokenStr string) (int64, string, time.Time, error) {
+func ValidateVerificationToken(tokenStr string) (int, string, time.Time, error) {
 	secret := os.Getenv("JWT_VERIFICATION_SECRET")
 	if secret == "" {
 		return 0, "", time.Time{}, errors.New("JWT_VERIFICATION_SECRET not set")
@@ -108,7 +108,7 @@ func ValidateVerificationToken(tokenStr string) (int64, string, time.Time, error
 	return claims.UserID, claims.Email, claims.VerificationExpiresAt, nil
 }
 
-func validateToken(tokenStr, secret string) (int64, error) {
+func validateToken(tokenStr, secret string) (int, error) {
 	if secret == "" {
 		return 0, errors.New("JWT secret not set")
 	}
